Modul deneme VM-V1/EdgeSolution/modules/ModbusClientModule/requirements.txt
azure-iot-device
schedule
pyModbusTCP

